Update Poll.js to allow selection of fourth option
When the fourth option is selected the whole string is passed to props instead of just the first character. This is because the first character was only being returned on the next iteration of the reduce function. By returning key[0], the selection of the first character occurs when we initially capture the vote of the user, and doesn't require an additional iteration.